requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM
ollama>=0.3.0
httpx>=0.25.0

# Embeddings
sentence-transformers>=3.0.0
torch>=2.0.0

# Vector Store
faiss-cpu>=1.7.0
qdrant-client>=1.7.0

# Document Processing
pymupdf>=1.23.0
markdown>=3.5.0

# Chunking
langchain-text-splitters>=0.3.0

# Prompts
jinja2>=3.1.0
pyyaml>=6.0.0

# Web Search
duckduckgo-search>=5.0.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# UI
streamlit>=1.30.0

# Monitoring
loguru>=0.7.0
prometheus-client>=0.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Dev
ruff>=0.2.0
